+++ src/E5R.Architecture.Core/Extensions/ObjectExtensions.cs
@@ -1,0 +1,20 @@
+// Copyright (c) E5R Development Team. All rights reserved.
+// This file is a part of E5R.Architecture.
+// Licensed under the Apache version 2.0: https://github.com/e5r/manifest/blob/master/license/APACHE-2.0.txt
+
+using System.Collections.Generic;
+
+namespace E5R.Architecture.Core.Extensions
+{
+    public static class ObjectExtensions
+    {
+        public static TObject Fill<TObject>(this TObject @object,
+            Dictionary<string, string> dictionary)
+        {
+            Checker.NotNullArgument(@object, nameof(@object));
+            Checker.NotNullArgument(dictionary, nameof(dictionary));
+
+            return dictionary.FillObject(@object);
+        }
+    }
+}